requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.2
requests==2.28.2
twilio==7.17.0
urllib3==1.26.15
